🗺️ Atlas: [architectural change] Extract router building logic from app.rs to router.rs
🕸️ Tangle: The `autumn/src/app.rs` file was significantly bloated (~2500 lines) because it was mixing the builder pattern for application configuration with the low-level complexities of constructing the Axum router and applying all framework middlewares, CORS, error pages, and static file serving logic.
📐 Blueprint: Extracted `build_router`, `build_router_merged`, `build_router_with_static`, and associated inner functions to a new `autumn/src/router.rs` module. Adjusted visibilities (`pub(crate)`) and updated call sites in `app.rs` and the test suite to use `crate::router::...`.
🧱 Stability: This reduces the coupling inside `app.rs`, keeping it focused solely on the user-facing builder API and startup flow.
🔬 Verification: `cargo clippy --all-targets --all-features` and `cargo test -p autumn-web` execute successfully with zero warnings.
