+++ autumn/tests/static_build_mode.rs
@@ -1,9 +1,9 @@
 //! Integration test: verify `render_static_routes` works with a real
 //! Autumn router (not just a mock fallback handler).
 
-use autumn_web::app::build_router;
 use autumn_web::config::AutumnConfig;
 use autumn_web::route::Route;
+use autumn_web::router::build_router;
 use autumn_web::static_gen::{StaticRouteMeta, render_static_routes};
 
 fn about_route() -> Route {
